+++ R/utils-grapes.R
@@ -7,7 +7,6 @@
 #' @examples
 #' 1:10 %nin% seq(from=2, to=10, by=2)
 #' c("a", "b", "c") %nin% c("a", "b")
-#' letters[letters %nin% unlist(strsplit("pack my box with five dozen liquor jugs", ""))]
 #' @export
 `%nin%` <- function (x, table) match(x, table, nomatch = 0) == 0